+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-sc
 
+## 2.4.21
+
+### Patch Changes
+
+- Bugfixes
+
+## 2.4.20
+
+### Patch Changes
+
+- Support multiple projects on the same domain
+
+## 2.4.19
+
+### Patch Changes
+
+- Sync package versions
+
 ## 1.5.6
 
 ### Patch Changes